+++ src/package/Support/Constants.php
@@ -13,4 +13,6 @@ class Constants
     const STATE_FAILED = 'failed';
 
     const STATE_RUNNING = 'running';
+
+    const CACHE_PROJECTS_KEY = 'tddd.projects.key';
 }